file: serverActions 폴더 위치 변경, SMS 인증 쿠키 이름 변경
